requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
firebase-admin==6.2.0
psycopg2-binary==2.9.9
asyncpg==0.28.0
geoalchemy2==0.14.1
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.0
websockets==12.0
passlib==1.7.4
alembic==1.12.1
pytest==7.4.3
email-validator==2.1.0
pillow==10.1.0
boto3==1.28.78
python-magic==0.4.27
aiofiles==23.2.1 